Rename form array helper to utils
